changed from using parse_priority to just Deserialize
I was already using serialize, might as well get on the bandwagon and
implement deserialize as well, it will be fine because my types are
allowing me to actually get what I need without needing to worry about
things like type bounds. This is where making an enum for the priority
was 100% the right decision because if this was just an int, or god
forbid, a string, then we would have had a harder time successfully
getting this into a valid Task format.